+++ account/models.py
@@ -8,6 +8,4 @@ class Account(models.Model):
     created_at      =       models.DateField(auto_now_add=True)
 
     class Meta:
-        db_table    =       "Accounts"
-
-
+        db_table    =       "accounts"
