+++ account/urls.py
@@ -4,8 +4,6 @@ from django.urls import path, include
 from .views import SignUpView, SignInView
 
 urlpatterns = [
-	path('signup', SignUpView.as_view()),
-	path('signin', SignInView.as_view())
+	path('sign-up', SignUpView.as_view()),
+	path('sign-in', SignInView.as_view())
 ]
-
-
